requirements.txt
streamlit>=1.32.0
plotly>=5.17.0
pandas>=2.2.0
numpy>=1.24.0
requests>=2.31.0
httpx>=0.27.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
scikit-learn>=1.4.0
statsmodels>=0.14.0
transformers>=4.41.0
torch>=2.2.0
web3>=6.15.0
arch>=6.3.0
tweepy>=4.14.0
prophet>=1.1.5
